move MockClientConnection into unit test utils module
should slightly reduce build time when cfg(not(test))
