+++ crates/valence_new/src/server/connection.rs
@@ -223,82 +223,3 @@ impl ClientConnection for RealClientConnection {
         }
     }
 }
-
-/// A mock client connection that can be used for testing.
-///
-/// Safe to clone, but note that the clone will share the same buffers.
-#[derive(Clone)]
-pub(crate) struct MockClientConnection {
-    buffers: Arc<Mutex<MockClientBuffers>>,
-}
-
-struct MockClientBuffers {
-    /// The queue of packets to receive from the client to be processed by the
-    /// server.
-    recv_buf: BytesMut,
-    /// The queue of packets to send from the server to the client.
-    send_buf: BytesMut,
-}
-
-impl MockClientConnection {
-    pub fn new() -> Self {
-        Self {
-            buffers: Arc::new(Mutex::new(MockClientBuffers {
-                recv_buf: BytesMut::new(),
-                send_buf: BytesMut::new(),
-            })),
-        }
-    }
-
-    pub fn inject_recv(&mut self, bytes: BytesMut) {
-        self.buffers.lock().unwrap().recv_buf.extend(bytes);
-    }
-
-    pub fn take_sent(&mut self) -> BytesMut {
-        self.buffers.lock().unwrap().send_buf.split()
-    }
-
-    pub fn clear_sent(&mut self) {
-        self.buffers.lock().unwrap().send_buf.clear();
-    }
-}
-
-impl ClientConnection for MockClientConnection {
-    fn try_send(&mut self, bytes: BytesMut) -> anyhow::Result<()> {
-        self.buffers.lock().unwrap().send_buf.extend(bytes);
-        Ok(())
-    }
-
-    fn try_recv(&mut self) -> anyhow::Result<BytesMut> {
-        Ok(self.buffers.lock().unwrap().recv_buf.split())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_mock_client_recv() -> anyhow::Result<()> {
-        let msg = 0xdeadbeefu32.to_be_bytes();
-        let b = BytesMut::from(&msg[..]);
-        let mut client = MockClientConnection::new();
-        client.inject_recv(b);
-        let b = client.try_recv()?;
-        assert_eq!(b, BytesMut::from(&msg[..]));
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_mock_client_send() -> anyhow::Result<()> {
-        let msg = 0xdeadbeefu32.to_be_bytes();
-        let b = BytesMut::from(&msg[..]);
-        let mut client = MockClientConnection::new();
-        client.try_send(b)?;
-        let b = client.take_sent();
-        assert_eq!(b, BytesMut::from(&msg[..]));
-
-        Ok(())
-    }
-}
